Fix "Elevation of privilege vulnerability in kernel ION subsystem"
CVE-2017-0507

An elevation of privilege vulnerability in the kernel ION subsystem could enable a local malicious application to execute arbitrary code within the context of the kernel. This issue is rated as Critical due to the possibility of a local permanent device compromise, which may require reflashing the operating system to repair the device.

# Conflicts:
#	drivers/staging/android/ion/ion.c