mt6735: Add an option to multiplex AP and STA on wlan0
This adds CONFIG_MTK_COMBO_AOSP_TETHERING_SUPPORT which, when enabled,
allows ap and wlan to co-exist in the same interface, as Android
expects.

Most of this functionality is also available (albeit not compilable
broken)
under CFG_TC1_FEATURE but that has larger implications around the radio
and usb stack that we do not want to adopt.

Change-Id: Ib1d1be40566f1bb9ccc7be45b49ec8d1f3b3ba58
Ticket: PORRIDGE-30

Signed-off-by: danielhk <daniel_hk>

# Conflicts:
#	drivers/misc/mediatek/connectivity/combo/drv_wlan/mt6630/wlan/os/linux/gl_init.c (reverted from commit 87de2a988b915dfa3a3aeb4dfd14c84526b66c4e)